+++ packages/config/eslint-preset.js
@@ -39,6 +39,8 @@ module.exports = {
         "packages/spor-card-react-native/",
         "packages/spor-icon-react-native/",
         "packages/spor-line-tag-react-native/",
+        "packages/spor-badge-react-native/",
+        "packages/spor-accordion-react-native/",
       ],
     },
   },