+++ packages/spor-accordion-react-native/src/ExpandableItem.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Box } from "@vygruppen/spor-layout-react-native";
+
+type ExpandableItemProps = {
+  children: React.ReactNode;
+};
+export const ExpandableItem = ({ children }: ExpandableItemProps) => {
+  return (
+    <Box marginBottom="sm" marginLeft="sm">
+      {children}
+    </Box>
+  );
+};