+++ packages/spor-theme-react-native/src/components/badge.tsx
@@ -1,0 +1,44 @@
+export const badgeColorSchemes = {
+  defaults: {
+    paddingLeft: 2,
+    paddingRight: 2,
+    borderRadius: "xl",
+    height: 24,
+    alignSelf: "center",
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  yellow: {
+    backgroundColor: "banana",
+    borderColor: "darkGrey",
+  },
+  "light-yellow": {
+    backgroundColor: "blonde",
+    borderColor: "golden",
+  },
+  green: {
+    backgroundColor: "seaMist",
+    borderColor: "darkTeal",
+  },
+  blue: {
+    backgroundColor: "lightBlue",
+    borderColor: "ocean",
+  },
+  grey: {
+    backgroundColor: "platinum",
+    borderColor: "darkGrey",
+  },
+  white: {
+    backgroundColor: "white",
+    borderColor: "silver",
+  },
+  orange: {
+    backgroundColor: "champagne",
+    borderColor: "pumpkin",
+  },
+  red: {
+    backgroundColor: "lightRed",
+    borderColor: "brightRed",
+  },
+};