+++ packages/spor-theme-react-native/src/components/index.tsx
@@ -2,3 +2,5 @@ export * from "./button";
 export * from "./card";
 export * from "./text";
 export * from "./lineicon";
+export * from "./badge";
+export * from "./expandable";