+++ packages/spor-card-react-native/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vygruppen/spor-card-react-native
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [02a0efb]
+  - @vygruppen/spor-theme-react-native@0.1.8
+
 ## 0.0.4
 
 ### Patch Changes